LOAD PRESET DETERMINES BEHAVIOUR OF ADC COMPUTATION

+++ macros/02AnaProcess.py
@@ -7,16 +7,17 @@ for run, ch in product(np.asarray(user_input["runs"]).astype(int),np.asarray(use
     ### Load
     my_runs = load_npy([run],[ch], info, preset=info["LOAD_PRESET"][2], compressed=True, debug=user_input["debug"])
     ### Process
-    compute_ana_wvfs(my_runs,info,debug=False)
-    delete_keys(my_runs,['RawADC','RawPeakAmp','RawPeakTime','RawPedSTD','RawPedMean','RawPedMax','RawPedMin','RawPedLim'])
+    if info["LOAD_PRESET"][2] == "ANA":
+        if check_key(my_runs[my_runs["NRun"][0]][my_runs["NChannel"][0]],"AnaADC") == False: compute_ana_wvfs(my_runs,info=info,debug=False)
+        delete_keys(my_runs,['RawADC','RawPeakAmp','RawPeakTime','RawPedSTD','RawPedMean','RawPedMax','RawPedMin','RawPedLim'])
     ### Compute
-    compute_peak_variables(my_runs, label="Ana", key="AnaADC", debug=user_input["debug"])
-    compute_pedestal_variables(my_runs, label="Ana", key="AnaADC", debug=user_input["debug"])
-    average_wvfs(my_runs, info=info, label="Ana", key="AnaADC", centering="NONE", debug=user_input["debug"])
-    average_wvfs(my_runs, info=info, label="Ana", key="AnaADC", centering="PEAK", debug=user_input["debug"])
-    compute_peak_RMS(my_runs, info=info, label="Ana", key="AnaADC", debug=user_input["debug"])
+    key, label = get_wvf_label(my_runs, "", "", debug=user_input["debug"])
+    compute_peak_variables(my_runs, info=info, key=key, label=label, debug=user_input["debug"])
+    compute_pedestal_variables(my_runs, info=info, key=key, label=label, debug=user_input["debug"])
+    average_wvfs(my_runs, info=info, key=key, label=label, centering="NONE", debug=user_input["debug"])
+    average_wvfs(my_runs, info=info, key=key, label=label, centering="PEAK", debug=user_input["debug"])
+    compute_peak_RMS(my_runs, info=info, key=key, label=label, debug=user_input["debug"])
     ### Remove branches to exclude from saving
-    # delete_keys(my_runs,["TimeStamp","Sampling"])
     save_proccesed_variables(my_runs,preset=str(info["SAVE_PRESET"][2]),info=info, force=True)
     ### Free memory
     del my_runs
